Move sudoku-solver-settings into folder sudoku-solver
