+++ src/app/components/sudoku-solver/sudoku-solver-settings/sudoku-solver-settings.component.scss
