+++ src/app/components/sudoku-solver/sudoku-solver.module.ts
@@ -4,10 +4,10 @@ import { IconModule } from "@app/components/icon/icon.module";
 import { NumberInputModule } from "@app/components/interactions/number-input/number-input.module";
 import { SectionModule } from "@app/components/section/section.module";
 import { SudokuGridModule } from "@app/components/sudoku-grid/sudoku-grid.module";
-import { SudokuSolverSettingsModule } from "@app/components/sudoku-solver-settings/sudoku-solver-settings.module";
 import { SudokuSolverEffects } from "@app/components/sudoku-solver/state/sudoku-solver.effects";
 import { SudokuSolverActionsComponent } from "@app/components/sudoku-solver/sudoku-solver-actions/sudoku-solver-actions.component";
 import { SudokuSolverBranchesComponent } from "@app/components/sudoku-solver/sudoku-solver-branches/sudoku-solver-branches.component";
+import { SudokuSolverSettingsModule } from "@app/components/sudoku-solver/sudoku-solver-settings/sudoku-solver-settings.module";
 import { SudokuSolverStatusComponent } from "@app/components/sudoku-solver/sudoku-solver-status/sudoku-solver-status.component";
 import { SudokuSolverStepsComponent } from "@app/components/sudoku-solver/sudoku-solver-steps/sudoku-solver-steps.component";
 import { SudokuSolverComponent } from "@app/components/sudoku-solver/sudoku-solver.component";
